Max. amplitude histo
Added a function to create a histogram storing the maximum amplitudes of the events of a channel.
I've just coppied the ChargeSpectrum function. Instead of filling the histo with the integral from windowind[1] to windowind[2], we fill it with the value of the maximum amplitude, that is stored in the bin windowind[0].
For the printing I've just coppied the PrintChargeSpectrum function, but changing the use of ChargeSpectrum for the new one MaxAmplitudeSpectrum.